requirements.txt
timm==0.9.10
tokenizers==0.19.1
transformers==4.40.1
accelerate==0.33.0
diffusers==0.29.2
flax==0.8.5
xformers
bitsandbytes
fastapi
uvicorn
json_numpy
draccus
numpy
pillow
requests
pandas
openpyxl
openai
torch>=2.2.0
torchvision>=0.16.0
